Decouple tool admin view-presenter
This is to more closely satisfy the `Supervising Controller` MVP pattern
described here; http://martinfowler.com/eaaDev/SupervisingPresenter.html

The primary part of this is the introduction of the event/handler
classes. Now, the presenter must register a handler to be informed of
the View's events, instead of the view having knowledge of the presenter
and making explicit calls to the presenter.

The second part of this is the assisted inject of the ListStore<> from
the presenter. Now, the presenter controls the list store, passing a
reference to the view so the view can display it to the user, and
communicate events related to it. The presenter updates the store
independently of the view.