Refactor ToolAdminDetails editor hierarchy
When you are working with editors, you only need one EditorDriver. There
were editor drivers sprinkled throughout the editor hierarchy which
starts at `ToolAdminDetailsView`.

Also set up DI for the classes in the hierarchy.

Finally, "data manipulation" which is needed to ensure that a service
call will succeed (like scrubbing `id` keys out of the beans) should be
kept at the service layer. Not at the view or presenter layer. So, that
is why some of the bean methods have been removed from the editors.